Limit the number of processes per blaster blastoff
This commit limits blaster to a certain amount of processes it can
spawn per blaster object. If the tasks passed to blaster exceeds more
than 10, it will only use 10 processes. Once a task is finished another
task will be handled by the process that just became free. If your
tasks is less than 10, it will spawn processes based on the number of
tasks to be blasted off.