stop processing tasks if one tasks fails when running sequentially
This commit brings the ability for blaster to stop processing any new
tasks when a task fails during sequential execution. When mode is set to
sequential, that means we want to proceed only if the previous tasks
completes successfully. When we run in parallel, we want everything to
run then process the results after the fact.